For VM names in Virtualbox to be identical to hostnames
VirtualBox 2.5.8 seems to have shortened VM name buffer to ~75
characters. Vagrant's automaticaly generated VM names don't seem to be
limited in length and can sometime exceed this limit.

Setting VM names to hostnames fix this and make VMs more recognizables
in the VirtualBox GUI but forbid to have multiple instances of the same
setup on a host.